lzbench.cpp: fix average time
I noticed that average compression/decompression time prints "0.00".

$ ~/github/lzbench/lzbench -p2 -i3,3 -elz4 -j *.hdr
lzbench 1.5 (64-bit Linux)   Assembled by P.Skibinski
Compressor name         Compress. Decompress. Compr. size  Ratio Filename
memcpy                   5301 MB/s  5317 MB/s   718722908 100.00 26345 files
lz4 1.7.3                0.00 MB/s  0.00 MB/s   369583316  51.42 26345 files
done... (cIters=3 dIters=3 cTime=1.0 dTime=2.0 chunkSize=1706MB cSpeed=0MB)

It turns out that the return value of std::accumulate is truncated
to int, because the initial value 0 has type int.  Changing it to
(uint64_t)0 solves the problem.

$ ~/github/lzbench/lzbench -p2 -i3,3 -elz4 -j *.hdr
lzbench 1.5 (64-bit Linux)   Assembled by P.Skibinski
Compressor name         Compress. Decompress. Compr. size  Ratio Filename
memcpy                   5202 MB/s  5266 MB/s   718722908 100.00 26345 files
lz4 1.7.3                 622 MB/s  2392 MB/s   369583316  51.42 26345 files
done... (cIters=3 dIters=3 cTime=1.0 dTime=2.0 chunkSize=1706MB cSpeed=0MB)

+++ _lzbench/lzbench.cpp
@@ -198,8 +198,8 @@ void print_stats(lzbench_params_t *params, const compressor_desc_t* desc, int le
             best_dtime = dtime[0];
             break;
         case AVERAGE: 
-            best_ctime = std::accumulate(ctime.begin(),ctime.end(),0) / ctime.size();
-            best_dtime = std::accumulate(dtime.begin(),dtime.end(),0) / dtime.size();
+            best_ctime = std::accumulate(ctime.begin(),ctime.end(),(uint64_t)0) / ctime.size();
+            best_dtime = std::accumulate(dtime.begin(),dtime.end(),(uint64_t)0) / dtime.size();
             break;
         case MEDIAN: 
             best_ctime = ctime[ctime.size()/2];
